Add ground floor in OverlayViewController
OverlayContainerController interprets its view controllers property differently.

If two or more view controllers are specified, the first view controller is used as a ground and the others ones are added as overlays over it.

If only one view controller is specified, the overlay container adds it as an single overlay without content. A backwards compatible behavior.